# Rewrite any and every in C