Reasons:

- they're very important, used in several places of bonus.stk and also in SRFIs
- it becomes incredibly faster and conses much less
- the C code is almost the same in size

Timings:

1. Fast path (single argument)
```scheme
(let ((X (iota 5000)))
  (time
    (repeat 10000 (any null? X))))

Old code:
2529.879 ms
Allocations: 1280000 bytes in 30000 allocation calls

New code:
1030.316 ms
Allocations: 0 bytes in 0 allocation calls
```

2. Normal path (two argument in this case)

```scheme
(let ((X (iota 5000))
      (Y  (iota 5000)))
  (time
    (repeat 10000 (every = X Y))))

Old code:
17490.161 ms
Allocations: 4801760000 bytes in 100050000 allocation calls

New code:
2991.282 ms
Allocations: 800960000 bytes in 50020000 allocation calls
```