@jpellegrini 2025-08-01
I also changed `boolean=?` in `r7rs.stk`.



The new version is:

```scheme
(define (boolean=? . args)
  (or (every (lambda (x) (eq? x #f)) args)
      (every (lambda (x) (eq? x #t)) args)))
```

With the new `every` this is also faster and conses less.

```scheme
(let ((a (map not (iota 1000))))
  (time
    (repeat 5000 (apply boolean=? a))))

Old code:
751.712 ms
120480000 bytes in 20000 allocation calls

new code, *without* the new every:
492.645 ms
121040000 bytes in 30000 allocation calls

new code, *with* the new every:
327.0 ms
120400000 bytes in 15000 allocation calls
```