@jpellegrini 2025-08-01
Now I realized:

* `symbol=?` can be optimized in the same way as `boolean=?`
* Both procedures would take at least two elements, according to R7RS.

So I included another patch in this PR addressing those issues.

For `symbol=?`:

```scheme
(let ((a (make-list 20000 'x)))
  (time
    (repeat 1000 (apply symbol=? a))))

Old code:
1786.216 ms
480096000 bytes in 4000 allocation calls

New code:
1276.569 ms
480080000 bytes in 3000 allocation calls
```