@jpellegrini 2025-08-11
> @egallesio -- I see in the errata of R7RS that " In fact it is an error to apply `boolean=?` to non-booleans". So we may leave it as it is (because R7RS lets us decide what to do with errors) or signal an error. Which would be preferable?

I have added a last patch to the PR, fixing that -- so if you prefer, you may ignore this last patch...