@egallesio 2025-08-21
> I have added a last patch to the PR, fixing that -- so if you prefer, you may ignore this last patch...

I have kept the test. I rewrote the function to be more efficient.